+++ examples/host/simpleResolver/test_simpleResolver.py
@@ -33,7 +33,7 @@ from openassetio.hostApi import ManagerFactory
 
 
 class Test_simpleResolver_errors:
-    def test_when_lib_initialization_fails_then_error_written_to_stderr_and_return_code_non_zero(
+    def test_when_no_config_then_error_written_to_stderr_and_return_code_non_zero(
         self, monkeypatch
     ):
         monkeypatch.delenv("OPENASSETIO_DEFAULT_CONFIG", raising=False)
@@ -41,6 +41,18 @@ class Test_simpleResolver_errors:
         expected_message = [
             "ERROR: No default manager configured, "
             f"check ${ManagerFactory.kDefaultManagerConfigEnvVarName}"
+        ]
+        assert result.stderr.splitlines() == expected_message
+        assert result.returncode == 1
+
+    def test_when_bad_config_then_error_written_to_stderr_and_return_code_non_zero(
+        self, monkeypatch
+    ):
+        monkeypatch.setenv("OPENASSETIO_DEFAULT_CONFIG", "/some/bad/path")
+        result = execute_cli()
+        expected_message = [
+            "ERROR: Could not load default manager config from '/some/bad/path', file does not"
+            " exist."
         ]
         assert result.stderr.splitlines() == expected_message
         assert result.returncode == 1